refactor: move tests into another folder, add overrides check in tests
